requirement.txt
astroid==1.6.6
backports.functools-lru-cache==1.6.1
configparser==4.0.2
Django==1.11.29
django-filter==2.0.0
djangorestframework==3.9.4
enum34==1.1.10
futures==3.3.0
isort==4.3.21
lazy-object-proxy==1.5.1
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.2.2
pkg-resources==0.0.0
pylint==1.9.5
pytz==2020.1
singledispatch==3.4.0.3
six==1.15.0
wrapt==1.12.1
